requirements.txt
asgiref==3.3.1
Django==3.1.4
django-tinymce==3.1.0
pytz==2020.4
sqlparse==0.4.1
